Turning initial try into bet of Elle and FranČesko
Elle thinks that despite random rewiring of the whole community of the dissatisfied agent (i.e., agent changes the whole network neighborhood), on the other hand FranČesko thinks that the network inevitably will have a random structure. We are not philosophers, we are scientists, so we ressolve our argument with the computational experiment!